# SchoolDiarySystem.WebClient/Api/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "SchoolDiarySystem.Api",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Subject": {
      "get": {
        "tags": [
          "Subject"
        ],
        "summary": "Получает все предметы.",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SubjectGetDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SubjectGetDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SubjectGetDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Subject"
        ],
        "summary": "Создает новый предмет.",
        "requestBody": {
          "description": "Данные для создания нового предмета.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubjectPostDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SubjectPostDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SubjectPostDto"
              }
            }
          }
        },
        "responses": {
          "400": {
            "description": "Если данные предмета некорректны.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "201": {
            "description": "Возвращает созданный предмет.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SubjectGetDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubjectGetDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubjectGetDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Subject/{id}": {
      "get": {
        "tags": [
          "Subject"
        ],
        "summary": "Получает предмет по идентификатору.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Идентификатор предмета.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Возвращает предмет.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SubjectGetDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubjectGetDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubjectGetDto"
                }
              }
            }
          },
          "404": {
            "description": "Если предмет не найден.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Subject"
        ],
        "summary": "Обновляет существующий предмет.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Идентификатор предмета для обновления.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "description": "Обновленные данные предмета.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubjectPostDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SubjectPostDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SubjectPostDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Возвращает обновленный предмет.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SubjectGetDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubjectGetDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubjectGetDto"
                }
              }
            }
          },
          "400": {
            "description": "Если данные предмета некорректны.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Если предмет не найден.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Subject"
        ],
        "summary": "Удаляет предмет по идентификатору.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Идентификатор предмета для удаления.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Возвращает сообщение об успешном удалении."
          },
          "404": {
            "description": "Если предмет не найден.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": {}
      },
...
      "SubjectGetDto": {
        "required": [
          "academicYear",
          "name",
          "subjectId"
        ],
        "type": "object",
        "properties": {
          "subjectId": {
            "type": "integer",
            "description": "Уникальный идентификатор предмета (первичный ключ).",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "description": "Название предмета.",
            "nullable": true
          },
          "academicYear": {
            "type": "integer",
            "description": "Год обучения, в котором преподается предмет.",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "SubjectPostDto": {
        "required": [
          "academicYear",
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Название предмета.",
            "nullable": true
          },
          "academicYear": {
            "type": "integer",
            "description": "Год обучения, в котором преподается предмет.",
            "format": "int32"
          }
        },
        "additionalProperties": false
      }
    }
  }
}